TF-instance and TF on curation form.
Let the curator select a TF and enter a TF accession.

If the TF_instance is already in DB, check entered TF value with the
TF-instance's TF value. If they are not same, return a error msg.

If the TF-instance is a new one, set TF-instance's TF to the value
entered by the curator.