Update resolved_refseq_to_uniprot.csv
Mapping complete. WP_018003289 (a ribosomal protein) needs to be looked up and see if it is used in any curation (does not show up in generated list).

YP_003250887 is temporarily mapped to C9RNT7 (an obsolete record), because the coding sequence for it has been annotated as pseudogene in Ensemble and there is no corresponding UniProt record (and no viable surrogate records).

A few of the automated mappings from each type should be manually verified.